seperate demodulators into files
